+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * providers should be imported and merged seperately (better treeshaking)
 * StorageType is now a proper enum
 * Universal might be working now
+* Update providers function
 
 #### 8.0.1
 
@@ -25,7 +26,6 @@
   * Using HttpClient instead of Http
   * Using HttpClient Interceptor instead of JwtHttp
   * In the `exchangeForToken` flow (when you want to exchange authorization code given by the authorization endpoint with your server's token) this library now sends a body that looks like this instead of a flat mesh of the oauth options, oauth result and user data:
-
 
 ```js
 {